+++ src/getStepStartStates.test.ts
@@ -37,6 +37,7 @@ import {
   STEP_RM_NETWORK_WATCHERS,
   STEP_RM_NETWORK_FLOW_LOGS,
   STEP_RM_NETWORK_LOCATION_WATCHERS,
+  STEP_RM_NETWORK_PRIVATE_ENDPOINTS,
 } from './steps/resource-manager/network';
 import { steps as storageSteps } from './steps/resource-manager/storage';
 import { IntegrationConfig } from './types';
@@ -121,6 +122,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_NETWORK_WATCHERS]: { disabled: true },
       [STEP_RM_NETWORK_LOCATION_WATCHERS]: { disabled: true },
       [STEP_RM_NETWORK_FLOW_LOGS]: { disabled: true },
+      [STEP_RM_NETWORK_PRIVATE_ENDPOINTS]: { disabled: true },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINE_IMAGES]: { disabled: true },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINE_DISKS]: { disabled: true },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINES]: { disabled: true },
@@ -219,6 +221,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_NETWORK_WATCHERS]: { disabled: true },
       [STEP_RM_NETWORK_LOCATION_WATCHERS]: { disabled: true },
       [STEP_RM_NETWORK_FLOW_LOGS]: { disabled: true },
+      [STEP_RM_NETWORK_PRIVATE_ENDPOINTS]: { disabled: true },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINE_IMAGES]: { disabled: true },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINE_DISKS]: { disabled: true },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINES]: { disabled: true },
@@ -317,6 +320,7 @@ describe('getStepStartStates', () => {
       [STEP_RM_NETWORK_WATCHERS]: { disabled: false },
       [STEP_RM_NETWORK_LOCATION_WATCHERS]: { disabled: false },
       [STEP_RM_NETWORK_FLOW_LOGS]: { disabled: false },
+      [STEP_RM_NETWORK_PRIVATE_ENDPOINTS]: { disabled: false },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINE_IMAGES]: { disabled: false },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINE_DISKS]: { disabled: false },
       [STEP_RM_COMPUTE_VIRTUAL_MACHINES]: { disabled: false },