+++ src/getStepStartStates.ts
@@ -41,6 +41,7 @@ import {
   STEP_RM_NETWORK_WATCHERS,
   STEP_RM_NETWORK_FLOW_LOGS,
   STEP_RM_NETWORK_LOCATION_WATCHERS,
+  STEP_RM_NETWORK_PRIVATE_ENDPOINTS,
 } from './steps/resource-manager/network/constants';
 import { steps as storageSteps } from './steps/resource-manager/storage/constants';
 import { STEP_RM_RESOURCES_RESOURCE_GROUPS } from './steps/resource-manager/resources/constants';
@@ -140,6 +141,7 @@ export function getResourceManagerSteps(): GetApiSteps {
       STEP_RM_COMPUTE_VIRTUAL_MACHINE_IMAGES,
       STEP_RM_COMPUTE_VIRTUAL_MACHINE_DISKS,
       STEP_RM_COMPUTE_VIRTUAL_MACHINES,
+      STEP_RM_NETWORK_PRIVATE_ENDPOINTS,
       computeSteps.VIRTUAL_MACHINE_EXTENSIONS,
       computeSteps.VIRTUAL_MACHINE_DISK_RELATIONSHIPS,
       STEP_RM_COSMOSDB_SQL_DATABASES,